Fix a few warnings related to integer precision

+++ Sources/ReaderViewController.m
@@ -60,7 +60,7 @@
 
 	BOOL isVisible;
 
-    int pagesVisible;
+    NSInteger pagesVisible;
 }
 
 #pragma mark Constants
